requirements.txt
albucore==0.0.13
albumentations==1.4.13
annotated-types==0.7.0
contourpy==1.2.1
cycler==0.12.1
deap==1.4.1
eval_type_backport==0.2.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
imageio==2.34.2
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mpmath==1.3.0
networkx==3.3
nn-torch==0.0.11
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.0.post2
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sympy==1.13.1
thop==0.1.1.post2209072238
tifffile==2024.7.24
tomli==2.0.1
torch==2.4.0
torch-optim==0.0.4
torch-pruning==1.4.2
torch-utils==0.1.2
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
wheel==0.44.0
